+++ openpecha/core/utils.py
@@ -1,0 +1,5 @@
+from uuid import uuid4
+
+
+def get_unique_id():
+    return uuid4().hex