+++ openpecha/formatters/__init__.py
@@ -1,3 +1,4 @@
+from .empty import EmptyEbook
 from .formatter import BaseFormatter
 from .google_orc import GoogleOCRFormatter
 from .hfml import HFMLFormatter