+++ tests/test_serialize.py
@@ -1,4 +1,4 @@
-from os import F_TLOCK
+import tempfile
 from pathlib import Path
 
 import pytest
@@ -18,16 +18,14 @@ def opf_path():
 
 @pytest.mark.skip(reason="not important")
 def test_hfml_serializer():
-
     opf_path = "./tests/data/serialize/hfml/P000002.opf/"
     output_path = "./output/P000002_hfml/"
     serializer = HFMLSerializer(opf_path)
     serializer.serialize(output_path=output_path)
 
 
-def test_opf_2_tsadra_serializer(opf_path):
-    serializer = EpubSerializer(opf_path)
-    serializer.apply_layers()
-    toc_levels = {}
-    out_fn = serializer.serialize(toc_levels, output_path="output/epub_output/")
-    assert str(out_fn) == "output/epub_output/P000100.epub"
+def test_hfml_2_tsadra_serializer(opf_path):
+    with tempfile.TemporaryDirectory() as tmpdirname:
+        serializer = EpubSerializer(opf_path)
+        out_fn = serializer.serialize(output_path=tmpdirname)
+        assert Path(out_fn).name == "P000100.epub"